monitoring/requirements.txt
Flask
requests
pyyaml
kubernetes
